pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "lxml>=6.0.2",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pypdf>=6.9.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "unstructured>=0.21.5",
    "wikipedia>=1.4.0",
]
